+++ CHANGELOG.md
@@ -5,3 +5,8 @@ Repository creation, with VERSION, README.md and CHANGELOG.md.
 # Version 1.0.0
 
 - Added all files for version 1.0.0
+
+# Version 2.0.0
+
+- Improved GUI for model and requirements setup 
+- Added artificial input signals and requirements